Set activate on single click to false for the language selection listbox
Resolves: https://github.com/sudipshil9862/fonts-compare/issues/43

We use listbox.set_selection_mode(Gtk.SelectionMode.SINGLE) that is
enough to select a language by a single click on a row. Using
listbox.set_activate_on_single_click(True) in addition to that causes
the problem that when an already selected row is clicked, the signal
“row-selected” is not emitted but the signal “row-activated” and we
did not connect anything to that signal. We also do not need to
connect a function to that signal because when we use

listbox.set_activate_on_single_click(False)

then a single click on the already selected row with the current
language will emit the signal “row-selected” and then everything works
as expected, the currently selected language is selected again, which
works just fine.